create descriptor_any_type
